Upgrade to use spring boot.
Modified the pom file, adding spring boot dependencies for web and jpa, and adjusting the existing dependencies to be compatible with it. The application now uses embedded Tomcat instead of Jetty, which is as standard. Formulated a new package structure, configured the spring boot application to find the Servlet, and performed a small refactoring to it, removing HibernateUtil.java and hibernate.cfg.xml along the way. Finally, added a new controller to serve the root jsp.